[main] The project stands for a bug-tracking platform. It allows for 2 types of users (Project Member and Tester) and they have different flows: testers for adding new bugs to the projects they enroll, and project members to manage the bugs reported and add new projects.
At the first launch, it is required to create a Project Member account so you can create a project. Otherwise, if you create a Tester account without any projects existing, you wasted your time: you have to disconnect and follow the steps above :).

In order to see the bugs/properties of a project, a user should be enrolled. By default, all the users existent at the moment of a project creation will be considered as testers/project members for that specific project. You can enroll by pressing the "Enroll" button on the Toolbar.

In order to disconnect from the app, you can easily do so by clicking the icon in the right-upper side of the app, on the Toolbar.